tests/requirements.txt
pytest>=7.4.0
pytest-xdist>=3.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
tomli>=2.0.0
